requirements.txt
Appium-Python-Client==1.2.0
autopep8==1.5.7
html-testRunner==1.2.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
selenium==3.141.0
toml==0.10.2
urllib3==1.26.6
